29.04.2025

+++ categories/CategoriesController.js
@@ -8,16 +8,22 @@ router.get("/admin/categories/new",(req, res) => {
 });
 
 router.post("/categories/save", (req, res) => {
-    var title = req.body.title;
-    if(title != undefined){
+    const title = req.body.title;
+    console.log("Titulo recebido:", title);
+
+
+    if(title != undefined && title.trim() !== "") {
         
         Category.create({
             title: title,
             slug: slugify(title)
-        }).then(() => {
-            res.redirect("/");
-        })
-
+        }).then(category => {
+            console.log("Categoria salva:", category);
+            res.redirect("/admin/categories");
+        }).catch(err => {
+            console.error("Erro ao salva:", err);
+            res.redirect("/admin/categories/new");
+        });
     }else{
         res.redirect("/admin/categories/new");
     }
@@ -25,27 +31,64 @@ router.post("/categories/save", (req, res) => {
 
 router.get("/admin/categories",  (req, res) => {
     Category.findAll().then(categories => {
+        console.log("Categorias encontradas:", categories);
         res.render("admin/categories/index", {categories: categories});
+    }).catch(err => {
+        console.error("Erro ao buscar categorias:", err);
+        res.redirect("/");
     });
 });
 
 router.post("/categories/delete", (req, res) => {
-    var id = req.body.id;
-    if(id != undefined){
-        if(!isNaN(id)){
+    const id = req.body.id;
+
+    if(id != undefined && !isNaN(id)) {
             Category.destroy({
-                where: {
-                    id: id
-                }
+                where: { id: id }
             }).then(() => {
+                console.log("Categoria deletada, ID:", id)
+                res.redirect("/admin/categories");
+            }).catch(err => {
+                console.error("Erro ao deletar categoria:", err);
                 res.redirect("/admin/categories");
             });
-        }else{// NÃO FOR UM NÚMERO
+        } else {// NÃO FOR UM NÚMERO
             res.redirect("/admin/categories");
         }
-    }else{ // NULL
-        res.redirect("/admin/categories");
-    }
 });
 
+// Rota para mostrar o formulário de edição
+router.get("/admin/categories/edit/:id", (req, res) => {
+    const id = req.params.id;
+
+    if (isNaN(id)) {
+        res.redirect("/admin/categories");
+    }
+
+    Category.findByPk(id).then(category => {
+        if (category != undefined) {
+            res.render("admin/categories/edit", { category: category });
+        } else {
+            res.redirect("/admin/categories");
+        }
+    }).catch(erro => {
+        res.redirect("/admin/categories");
+    });
+});
+
+// Rota para atualizar no banco
+router.post("/categories/update", (req, res) => {
+    const id = req.body.id;
+    const title = req.body.title;
+
+    Category.update(
+        { title: title, slug: slugify(title) },
+        { where: { id: id } }
+    ).then(() => {
+        res.redirect("/admin/categories");
+    }).catch((err) => {
+        console.error("Erro ao atualizar categoria:", err);
+        res.redirect("/admin/categories");
+    });
+});
 module.exports = router;
